requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Graph & Data
pandas
networkx

# Web Scraping
selenium
undetected-chromedriver
beautifulsoup4

# Utilities
setuptools
python-dotenv
